+++ Cli-CredentialHelper/Cli-CredentialHelper.csproj
@@ -78,6 +78,10 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\GitHub.Authentication\GitHub.Authentication.csproj">
+      <Project>{cf306116-bbf0-4cc7-afce-a506ac4752cb}</Project>
+      <Name>GitHub.Authentication</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Microsoft.Alm.Authentication\Microsoft.Alm.Authentication.csproj">
       <Project>{c1daabc1-b493-459d-bb4f-04fbefb1ba13}</Project>
       <Name>Microsoft.Alm.Authentication</Name>
