+++ Microsoft.Alm.Authentication/GitHubAuthentication.cs
@@ -1,10 +1,6 @@
 ﻿using System;
 using System.Diagnostics;
-using System.Threading;
 using System.Threading.Tasks;
-using System.Windows;
-using GitHub.Authentication.ViewModels;
-using GitHub_Authentication;
 
 namespace Microsoft.Alm.Authentication
 {
@@ -142,6 +138,7 @@ namespace Microsoft.Alm.Authentication
         }
 
         /// <summary>
+        /// <para></para>
         /// <para>Tokens acquired are stored in the secure secret store provided during
         /// initialization.</para>
         /// </summary>
@@ -313,58 +310,5 @@ namespace Microsoft.Alm.Authentication
         /// </param>
         /// <param name="result">The result of the interactive authenticaiton attempt.</param>
         public delegate void AuthenticationResultDelegate(TargetUri targetUri, GithubAuthenticationResultType result);
-
-        public static bool GithubAuthcodeModalPrompt(TargetUri targetUri, GithubAuthenticationResultType resultType, string username, out string authenticationCode)
-        {
-            Trace.WriteLine("Program::GithubAuthcodeModalPrompt");
-
-            var twoFactorViewModel = new TwoFactorViewModel(resultType == GithubAuthenticationResultType.TwoFactorSms);
-
-            Trace.WriteLine("   prompting user for authentication code.");
-
-            StartSTATask(() =>
-            {
-                if (!UriParser.IsKnownScheme("pack"))
-                {
-                    UriParser.Register(new GenericUriParser(GenericUriParserOptions.GenericAuthority), "pack", -1);
-                }
-                var app = new Application();
-                var appResources = new Uri("pack://application:,,,/GitHub.Authentication;component/AppResources.xaml", UriKind.RelativeOrAbsolute);
-                app.Resources.MergedDictionaries.Add(new ResourceDictionary { Source = appResources });
-                app.Run(new TwoFactorWindow { ViewModel = twoFactorViewModel });
-            })
-            .Wait();
-
-            // If the user cancels the dialog, we need to ignore anything they've
-            // typed into the authentication code.
-            bool authenticationCodeValid = 
-                twoFactorViewModel.Result == TwoFactorResult.Ok
-                && twoFactorViewModel.IsValid;
-            
-            authenticationCode = authenticationCodeValid
-                ? twoFactorViewModel.AuthenticationCode
-                : null;
-            return authenticationCodeValid;
-        }
-
-        static Task StartSTATask(Action action)
-        {
-            var completionSource = new TaskCompletionSource<object>();
-            var thread = new Thread(() =>
-            {
-                try
-                {
-                    action();
-                    completionSource.SetResult(null);
-                }
-                catch (Exception e)
-                {
-                    completionSource.SetException(e);
-                }
-            });
-            thread.SetApartmentState(ApartmentState.STA);
-            thread.Start();
-            return completionSource.Task;
-        }
     }
 }
